+++ packages/visual-editor/src/store/entityStore.ts
@@ -28,6 +28,9 @@ export const useEntityStore = create<EntityState>((set, get) => ({
     set({ areEntitesResolvedInParent: resolved });
   },
   resetEntityStore(locale, entities = []) {
+    console.debug(
+      `[exp-builder.sdk] Resetting entity store because the locale changed to '${locale}'.`
+    );
     set({ entityStore: new EditorModeEntityStore({ locale, entities }) });
   },
 }));